+++ lib/Conversion/TritonGPUToLLVM/TypeConverter.h
@@ -8,6 +8,20 @@
 using namespace mlir;
 using namespace mlir::triton;
 
+// Copied from GpuOpsLowering.h
+/// A function that maps a MemorySpace enum to a target-specific integer value.
+using MemorySpaceMapping =
+    std::function<unsigned(mlir::gpu::AddressSpace gpuAddressSpace)>;
+
+/// Copied from GpuOpsLowering.cpp
+static IntegerAttr wrapNumericMemorySpace(MLIRContext *ctx, unsigned space) {
+  return IntegerAttr::get(IntegerType::get(ctx, 64), space);
+}
+
+// Copied from GpuOpsLowering.cpp
+void populateGpuMemorySpaceAttributeConversions(
+    TypeConverter &typeConverter, const MemorySpaceMapping &mapping);
+
 class TritonGPUToLLVMTypeConverter : public LLVMTypeConverter {
 public:
   using TypeConverter::convertType;
